Add a filter for final ostrichcized paths.
Additionally, change language so we have directories (first part of a path), plugins (the slug for the plugin that forms the end of the path), and the paths (the full path including the directory and the plugin slug). With being able to filter the final paths, one can granularly control all paths that will be excluded from error reports.