Better handling of MU plugin scenarios.
In order to catch and hide more errors, the ostrich should be initiated on 'muplugins_loaded' when the plugin is used as an MU plugin. This will catch more plugin issues related to authors not running code on a hooks and filters and instead running code directly in the .php file. Issue #8 clearly documents these concerns.